+++ ui/shared/components/panel/search/constants.js
@@ -478,7 +478,6 @@ export const IN_SILICO_FIELDS = [REQUIRE_SCORE_FIELD, ...PREDICTOR_FIELDS.filter
     )
     return {
       labelHelp,
-      popupProps: thresholds ? { hoverable: true } : {},
       control: Form.Input,
       type: 'number',
       min: min || 0,
